feat: moved scan & qr components
