+++ src/renderer/entities/transaction/ui/QrCode/QrGenerator/QrDerivationsGenerator.tsx
