+++ src/renderer/entities/transaction/ui/QrCode/QrGenerator/QrTxGenerator.test.tsx
