+++ src/renderer/entities/transaction/ui/QrCode/QrGenerator/QrTxGenerator.tsx
