+++ src/renderer/entities/transaction/ui/QrCode/QrGenerator/common/constants.ts
