+++ src/renderer/entities/transaction/ui/QrCode/QrGenerator/common/useGenerator.ts
