+++ src/renderer/entities/transaction/ui/QrCode/QrGenerator/common/utils.ts
