+++ src/renderer/entities/transaction/ui/QrCode/QrReader/QrReader.test.tsx
@@ -3,7 +3,7 @@ import { BrowserCodeReader, BrowserQRCodeReader } from '@zxing/browser';
 
 import { QR_READER_ERRORS } from '../common/errors';
 import { QrError } from '../common/types';
-import QrReader from './QrReader';
+import { QrReader } from './QrReader';
 
 jest.mock('raptorq');
