+++ src/renderer/entities/transaction/ui/QrCode/QrReader/RaptorFrame.ts
