+++ src/renderer/entities/transaction/ui/QrCode/QrReader/style.css
