+++ src/renderer/entities/transaction/ui/QrCode/common/constants.ts
