+++ src/renderer/entities/transaction/ui/QrCode/common/errors.ts
