+++ src/renderer/entities/transaction/ui/QrCode/common/types.ts
