+++ src/renderer/entities/transaction/ui/index.ts
@@ -1,10 +1,16 @@
-import { Deposit } from './Deposit/Deposit';
-import { DepositWithLabel } from './DepositWithLabel/DepositWithLabel';
-import { Fee } from './Fee/Fee';
-import { OperationResult } from './OperationResult/OperationResult';
-
-export * from './OperationResult/common/constants';
-export type { Variant } from './OperationResult/common/types';
-
-export { Fee, DepositWithLabel, OperationResult, Deposit };
+export { Deposit } from './Deposit/Deposit';
+export { DepositWithLabel } from './DepositWithLabel/DepositWithLabel';
+export { Fee } from './Fee/Fee';
 export { XcmFee } from './XcmFee/XcmFee';
+export { OperationResult } from './OperationResult/OperationResult';
+export { ScanSingleframeQr } from './Scanning/ScanSingleframeQr';
+export { ScanMultiframeQr } from './Scanning/ScanMultiframeQr';
+
+// TODO: requires refactoring clickup task - https://app.clickup.com/t/86933e82e
+export { cryptoTypeToMultisignerIndex } from './QrCode/QrGenerator/common/utils';
+export { QrDerivationsGenerator } from './QrCode/QrGenerator/QrDerivationsGenerator';
+export { QrTextGenerator } from './QrCode/QrGenerator/QrTextGenerator';
+export { QrReader } from './QrCode/QrReader/QrReader';
+export { QrReaderWrapper } from './QrCode/QrReader/QrReaderWrapper';
+export * from './QrCode/common/types';
+export * from './QrCode/common/constants';
