+++ src/renderer/pages/Notifications/Notifications.tsx
@@ -4,9 +4,8 @@ import { format } from 'date-fns';
 import { useI18n } from '@app/providers';
 import EmptyNotifications from './components/EmptyNotifications';
 import { sortByDate } from './common/utils';
-import { FootnoteText } from '@shared/ui';
+import { FootnoteText, Header } from '@shared/ui';
 import { useNotification, NotificationRow } from '@entities/notification';
-import { Header } from '@renderer/components/common';
 
 export const Notifications = () => {
   const { t, dateLocale } = useI18n();
