+++ src/renderer/pages/Onboarding/Vault/ManageVault/model/manage-vault-model.ts
@@ -2,13 +2,13 @@ import { createApi, createEvent, createStore, sample, combine, createEffect, att
 import { createForm } from 'effector-forms';
 import { spread } from 'patronum';
 
-import { SeedInfo } from '@renderer/components/common/QrCode/common/types';
 import { chainsService } from '@entities/network';
 import { accountUtils, KEY_NAMES, walletModel } from '@entities/wallet';
 import type { ChainAccount, ShardAccount, DraftAccount, BaseAccount, Wallet, Account, NoID } from '@shared/core';
 import { AccountType, ChainType, CryptoType, KeyType } from '@shared/core';
 import { dictionary } from '@shared/lib/utils';
 import { storageService } from '@shared/api/storage';
+import { SeedInfo } from '@entities/transaction';
 
 const chains = chainsService.getChainsData();
