+++ src/renderer/pages/Onboarding/Vault/ScanStep/ScanStep.tsx
@@ -1,9 +1,9 @@
-import { SeedInfo } from '@renderer/components/common/QrCode/common/types';
 import { useI18n } from '@app/providers';
 import KeyQrReader from '../KeyQrReader/KeyQrReader';
 import { Button, HeaderTitleText, SmallTitleText } from '@shared/ui';
 import onboarding_tutorial from '@shared/assets/video/onboarding_tutorial.mp4';
 import onboarding_tutorial_webm from '@shared/assets/video/onboarding_tutorial.webm';
+import { SeedInfo } from '@entities/transaction';
 
 type Props = {
   onBack: () => void;
