+++ src/renderer/pages/Onboarding/Vault/Vault.tsx
@@ -1,6 +1,5 @@
 import { useEffect, useState } from 'react';
 
-import { SeedInfo } from '@renderer/components/common/QrCode/common/types';
 import ScanStep from './ScanStep/ScanStep';
 import { ManageMultishard } from './ManageMultishard/ManageMultishard';
 import { ManageSingleshard } from './ManageSingleshard/ManageSingleshard';
@@ -8,7 +7,7 @@ import { ManageVault } from './ManageVault/ManageVault';
 import { BaseModal } from '@shared/ui';
 import { DEFAULT_TRANSITION } from '@shared/lib/utils';
 import { useToggle } from '@shared/lib/hooks';
-import { VaultFeatures } from '@renderer/components/common/QrCode/common/constants';
+import { SeedInfo, VaultFeatures } from '@entities/transaction';
 
 const isDynamicDerivationSupport = (seedInfo: SeedInfo): boolean => {
   const dynamicDerivationsExist = seedInfo.features?.some(
