+++ src/renderer/pages/Operations/Operations.tsx
@@ -7,10 +7,9 @@ import { useI18n } from '@app/providers';
 import EmptyOperations from './components/EmptyState/EmptyOperations';
 import Operation from './components/Operation';
 import { sortByDateDesc } from './common/utils';
-import { FootnoteText } from '@shared/ui';
+import { FootnoteText, Header } from '@shared/ui';
 import { MultisigTransactionDS } from '@shared/api/storage';
 import { useMultisigTx, useMultisigEvent } from '@entities/multisig';
-import { Header } from '@renderer/components/common';
 import { MultisigEvent, MultisigTransactionKey } from '@entities/transaction';
 import { OperationsFilter } from '@features/operation';
 import { walletModel, accountUtils } from '@entities/wallet';
