+++ src/renderer/shared/ui/Header/Header.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 
-import Header from './Header';
+import { Header } from './Header';
 
 describe('components/common/Header', () => {
   test('should render component', () => {
