+++ src/renderer/widgets/ReceiveAssetModal/ui/ReceiveAssetModal.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { useUnit } from 'effector-react';
 
-import { OperationTitle, QrTextGenerator } from '@renderer/components/common';
+import { OperationTitle } from '@renderer/components/common';
 import { DefaultExplorer, ExplorerIcons } from '@shared/ui/ExplorerLink/constants';
 import { BaseModal, Button, FootnoteText, HelpText, Icon, Select } from '@shared/ui';
 import { DropdownOption, DropdownResult } from '@shared/ui/types';
@@ -10,6 +10,7 @@ import { copyToClipboard, DEFAULT_TRANSITION, toAddress, cnTw } from '@shared/li
 import { AccountAddress, walletModel, walletUtils, accountUtils } from '@entities/wallet';
 import { useToggle } from '@shared/lib/hooks';
 import type { Chain, Asset } from '@shared/core';
+import { QrTextGenerator } from '@entities/transaction';
 
 type Props = {
   chain: Chain;
